sql: extract distsql check code into a separate file
This commit extracts the code for checking distsql supportability into
a separate file. This is done because the following commit will be
extending the state we're storing during the distsql check.

It's purely a mechanical move, with a couple of tweaks:
- `checkSupportForInvertedFilterNode` is inlined in its single call
site (to get symmetry with all other planNodes).
- all unsupported errors have been extracted into global singletons. The
following commit will take advantage of this by replacing errors with
bit-map of causes - the loss of some extra context should be negligible.

Release note: None
